[iOS] fix : 개인 기록 가져올때 데이터 포맷 달라 나는 오류 해결
#124

+++ iOS/EatUpAll/EatUpAll/Home/Models/PersonalTotalSaving.swift
@@ -11,5 +11,5 @@ import Foundation
 struct PersonalTotalSaving: Codable {
     private(set) var accountName: String
     private(set) var totalPlates: Int
-    private(set) var totalSaving: Int
+    private(set) var totalSaving: Double
 }
